mTrc substantially modified
The execution trace is now written to a log file which defaults to "ExecTrace.log" in ThisWorkbook's parent folder. By default the file is new with each trace begin. The property LogFile not only provided a means to specify another log file at whichever location but also specify that logs are appended.